+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 ## [Unreleased]
 
+### Added
+
+- [[192]](https://github.com/rust-vmm/vm-memory/pull/192): Add new method
+  as_any() to GuestAddressSpace
+
 ## [v0.8.0]
 
 ### Fixed